+++ src/components/layout/mobile/aside.vue
@@ -21,7 +21,7 @@
     </div>
     <div class="aside-nav">
       <nav class="nav-list">
-        <router-link to="/" class="item" exact>
+        <router-link to="/" class="item">
           <i class="iconfont icon-home-fill"></i>
           <i18n :k="LanguageKey.PAGE_HOME" />
         </router-link>