+++ src/config/app.config.ts
@@ -20,7 +20,7 @@ export const META = Object.freeze({
 })
 
 export const THIRD_IDS = Object.freeze({
-  YOUTUBE_CHANNEL_ID: `UCoL-j6T28PLSJ2U6ZdONS0w`,
+  YOUTUBE_CHANNEL_ID: 'UCoL-j6T28PLSJ2U6ZdONS0w',
   MUSIC_163_BGM_ALBUM_ID: '638949385',
   GITHUB_USER_ID: 'surmon-china',
   TWITTER_USER_ID: 'surmon7788',
@@ -56,6 +56,7 @@ export const VALUABLE_LINKS = Object.freeze({
   MUSIC_163: `https://music.163.com/#/playlist?id=${THIRD_IDS.MUSIC_163_BGM_ALBUM_ID}`,
   YOUTUBE_CHANNEL: `https://www.youtube.com/channel/${THIRD_IDS.YOUTUBE_CHANNEL_ID}`,
   TELEGRAM: 'https://t.me/surmon',
+  OPENSEA: 'https://opensea.io/surmon',
   DOUBAN: 'https://www.douban.com/people/nocower',
   ZHIHU: 'https://www.zhihu.com/people/surmon',
   QUORA: 'https://www.quora.com/profile/Surmon',